+++ src/__test__/Function.test.js
@@ -1,0 +1,5 @@
+import { getCountriesApi } from '../redux/countries';
+
+it('getCountriesApi should return a promise', () => {
+  expect(getCountriesApi()).toEqual(expect.any(Function));
+});
