+++ src/__test__/Searched.test.js
@@ -1,0 +1,5 @@
+import { getSearchedCountries } from '../redux/countries';
+
+it('getSearchedCountries should return a promise', () => {
+  expect(getSearchedCountries()).toEqual(expect.any(Function));
+});
